pyproject.toml
[project]
name = "gmail-assistant"
version = "0.1.0"
description = "MCP server providing Gmail utility tools (mark-as-read, fetch body) for Claude Code"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.2.0",
    "google-api-python-client>=2.193.0",
    "google-auth>=2.49.1",
    "google-auth-oauthlib>=1.3.1",
    "markdownify>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["mcp"]
